src/main/java/com/example/demo/flight/FlightController.java
package com.example.demo.flight;

import com.example.demo.model.Flight;
import com.example.demo.service.FlightService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/flights")
public class FlightController {

    @Autowired
    FlightService flightService;

    @GetMapping("")
    public List<Flight> flights() {
        return flightService.getFlights();
    }

    @GetMapping("/flight")
    public Flight flight() {
        return flightService.getFlight();
    }

}
